+++ custom_components/flexdisplay/const.py
@@ -10,6 +10,7 @@ PLATFORMS = [
     "binary_sensor",
     "button",
     "event",
+    "image",
     "select",
     "number",
     "switch",